Add files via upload
Contact Us Page
I re-uploaded the contactus page, something in the CSS was weird and the titles were yellow, so I just made that local as a bandaid for the moment
Fixed a rogue ">" at the top of the screen lol

New Files
Started a individualbook.html file for each individual book where you can get more information on it, like a quick synopsis and add to cart button, etc.  Will need to connect this to database later.  PS, can our database support description text?  It's kind of long
Added a new folder for pictures of  random books under "images"