Slightly breaks and fixes
Added a current books nav in admin and linked it to the DB. This now links to index.html as well so that if you have a book in the DB it will allow you to click the book name and brings you to its individual HTML page. As well, I broke it. It currently does not have the right name generation since it is trying to base it off of an ID (ie. 12-TEST.html instead of TEST.html).